Make initialize access-only.
