+++ tests/test_download.py
@@ -96,78 +96,6 @@ def test_empty_payload(client):
     assert consume.status_code == 400
 
 
-@pytest.mark.integration
-def test_initialize_on_bad_url(client, publisher_wallet, consumer_wallet, web3):
-    asset = get_dataset_with_invalid_url_ddo(client, publisher_wallet)
-    service = get_first_service_by_type(asset, ServiceType.ACCESS)
-
-    mint_100_datatokens(
-        web3, service.datatoken_address, consumer_wallet.address, publisher_wallet
-    )
-
-    response = initialize_service(
-        client, asset.did, service, consumer_wallet, raw_response=True
-    )
-    assert "error" in response.json
-    assert response.json["error"] == "Asset URL not found or not available."
-
-
-@pytest.mark.integration
-def test_initialize_on_ipfs_url(client, publisher_wallet, consumer_wallet, web3):
-    asset = get_dataset_with_ipfs_url_ddo(client, publisher_wallet)
-    service = get_first_service_by_type(asset, ServiceType.ACCESS)
-
-    mint_100_datatokens(
-        web3, service.datatoken_address, consumer_wallet.address, publisher_wallet
-    )
-
-    datatoken, nonce, computeAddress, providerFees = initialize_service(
-        client, asset.did, service, consumer_wallet
-    )
-
-    assert datatoken == service.datatoken_address
-
-
-@pytest.mark.integration
-def test_initialize_on_disabled_asset(client, publisher_wallet, consumer_wallet, web3):
-    asset, real_asset = get_dataset_ddo_disabled(client, publisher_wallet)
-    assert real_asset
-    service = get_first_service_by_type(asset, ServiceType.ACCESS)
-
-    mint_100_datatokens(
-        web3, service.datatoken_address, consumer_wallet.address, publisher_wallet
-    )
-
-    response = initialize_service(
-        client, asset.did, service, consumer_wallet, raw_response=True
-    )
-    assert "error" in response.json
-    assert response.json["error"] == "Asset is not consumable."
-
-
-@pytest.mark.integration
-def test_initialize_on_asset_with_custom_credentials(
-    client, publisher_wallet, consumer_wallet, web3
-):
-    asset = get_dataset_ddo_with_denied_consumer(
-        client, publisher_wallet, consumer_wallet.address
-    )
-
-    service = get_first_service_by_type(asset, ServiceType.ACCESS)
-
-    mint_100_datatokens(
-        web3, service.datatoken_address, consumer_wallet.address, publisher_wallet
-    )
-
-    response = initialize_service(
-        client, asset.did, service, consumer_wallet, raw_response=True
-    )
-    assert "error" in response.json
-    assert (
-        response.json["error"]
-        == f"Error: Access to asset {asset.did} was denied with code: ConsumableCodes.CREDENTIAL_IN_DENY_LIST."
-    )
-
 
 @pytest.mark.integration
 def test_download_multiple_files(client, publisher_wallet, consumer_wallet, web3):
